runable tf-agents with model built

+++ colab/36_01_tf_agents_qhuy_phm/miner_env_9x21x2.py
@@ -87,7 +87,8 @@ class MinerEnv:
         else:
             channel1[self.state.y, self.state.x] = self.state.energy
         state = np.stack((channel0, channel1), axis=-1)
-        return state.astype(np.int32)
+        #return state.astype(np.int32)
+        return state.astype(np.float32)
 
 
     def get_reward(self):
